refactor and add dynamodb
